Display Filter - Conditional Feature Display on Maps
Do you want to keep access to all features in a layer but want to display just a few on them on your map? In ArcGIS Pro we can achieve this with Display Filters. This functionality allow us to use queries that limit which features of a layer are displayed on determined scales. On the example below, we are using different scales to display buildings according to its size.  
Display Filters differ from Definition Queries. Display filters affect the map display only, it means that all the features are still available for query, analysis, edition and geoprocessing operations. When we filter features using the Definition Queries functionality, filtered features are not available to the layer at all.
This is a handy tool to apply conditional scale-based display on your maps. This can help improving the user experience of basemaps and web map applications.

#gis #cartography #mapdesign #arcgispro #webmapapplication